backend/requirements.txt
aioredis==1.3.1
asgiref==3.2.3
astroid==2.3.3
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.2.2
Automat==20.2.0
autopep8==1.5
backports.csv==1.0.7
certifi==2019.11.28
cffi==1.14.0
channels==2.4.0
channels-redis==2.4.2
chardet==3.0.4
Click==7.0
constantly==15.1.0
cryptography==2.9
daphne==2.5.0
defusedxml==0.6.0
diff-match-patch==20181111
dj-database-url==0.5.0
Django==3.0.5
django-allauth==0.41.0
django-channels-presence==1.0.0
django-cors-headers==3.2.1
django-error-report==0.2.0
django-import-export==2.0.2
django-rest-passwordreset==1.1.0
djangorestframework==3.11.0
et-xmlfile==1.0.1
hiredis==1.0.1
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
jdcal==1.4.1
lazy-object-proxy==1.4.3
MarkupPy==1.14
msgpack==0.6.2
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.3
Pillow==7.1.1
psycopg2-binary==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycparser==2.19
PyHamcrest==2.0.1
PyJWT==1.7.1
pyOpenSSL==19.1.0
python-social-auth==0.3.6
python3-openid==3.1.0
pytz==2019.3
PyYAML==5.3
requests==2.23.0
requests-oauthlib==1.3.0
service-identity==18.1.0
six==1.14.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
tablib==1.1.0
Twisted==19.10.0
txaio==20.1.1
urllib3==1.25.9
websockets==8.1
whitenoise==5.0.1
wrapt==1.11.2
xlrd==1.2.0
xlwt==1.3.0
zope.interface==5.1.0
django_storages==1.9.1
google-cloud-storage==1.28.1
celery==4.4.2
django-celery-beat==2.0.0
redis==3.5.2
numpy==1.21.2
opencv_python_headless==4.5.3.56